requirements.txt
PyYAML>=5.3.1
discord.py>=1.5.1